requirements.txt
Django==3.2.8
dnspython3==1.15.0
djangorestframework==3.12.4
gunicorn==20.1.0